Changed delimiter to the colon character.
Since we're not currently escaping the delim
in object keys the delim should be something
people wouldn't use as object keys.

Since the colon char is already used to separate
object keys/vals it's not asking too much for
users of the format to not use the char in their
object keys.

Of course, the RFC couldn't care less if colons
appeared in object keys.

Gotta start somewhere...